Move scipy wrappers to subpackage
