+++ src/scipp/scipy/interpolate/__init__.py
@@ -7,9 +7,9 @@ This subpackage provides wrappers for a subset of functions from
 :py:mod:`scipy.interpolate`.
 """
 
-from ..core import empty, epoch, Variable, DataArray, DimensionError, UnitError
-from ..core import DType, irreducible_mask
-from ..compat.wrapping import wrap1d
+from ...core import empty, epoch, Variable, DataArray, DimensionError, UnitError
+from ...core import DType, irreducible_mask
+from ...compat.wrapping import wrap1d
 
 from typing import Any, Callable, Literal, Union
